fix: ビルドエラーとlintエラーを修正
