+++ apps/web/src/app/api/internal/metrics/collect/route.ts
@@ -1,15 +1,15 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { metricsHelpers } from '@/app/api/metrics/route'
+import { metricsHelpers } from '@/lib/metrics'
 
 export async function POST(request: NextRequest) {
   try {
     const data = await request.json()
     const { method, route, statusCode, duration } = data
-    
+
     // メトリクスを記録
     metricsHelpers.incrementHttpRequests(method, route, statusCode)
     metricsHelpers.recordHttpDuration(method, route, statusCode, duration)
-    
+
     return NextResponse.json({ success: true })
   } catch (error) {
     console.error('Error collecting metrics:', error)
