+++ apps/web/src/middleware.ts
@@ -1,16 +1,15 @@
 import { NextRequest, NextResponse } from 'next/server'
-import { withAuth } from 'next-auth/middleware'
 
 // メトリクス収集用のヘルパー
 async function collectMetrics(request: NextRequest, response: NextResponse) {
   const startTime = Date.now()
   const method = request.method
   const pathname = request.nextUrl.pathname
-  
+
   // レスポンスのステータスコードを取得（デフォルトは200）
   const statusCode = response.status || 200
   const duration = (Date.now() - startTime) / 1000 // 秒単位
-  
+
   // メトリクスAPIにデータを送信（非同期で実行）
   try {
     const metricsData = {
@@ -19,7 +18,7 @@ async function collectMetrics(request: NextRequest, response: NextResponse) {
       statusCode,
       duration,
     }
-    
+
     // メトリクス収集は失敗してもアプリケーションに影響しないようにする
     fetch(`${request.nextUrl.origin}/api/internal/metrics/collect`, {
       method: 'POST',
@@ -35,17 +34,20 @@ async function collectMetrics(request: NextRequest, response: NextResponse) {
 
 export function middleware(request: NextRequest) {
   const pathname = request.nextUrl.pathname
-  
+
   // メトリクスエンドポイント自体は計測しない
-  if (pathname === '/api/metrics' || pathname === '/api/internal/metrics/collect') {
+  if (
+    pathname === '/api/metrics' ||
+    pathname === '/api/internal/metrics/collect'
+  ) {
     return NextResponse.next()
   }
-  
+
   const response = NextResponse.next()
-  
+
   // メトリクス収集（非同期）
   collectMetrics(request, response)
-  
+
   return response
 }
